% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResultsSim.R
\name{getResultsSim}
\alias{getResultsSim}
\title{getResults}
\usage{
getResultsSim(VarSelectObject, numGroups = 2, cuts = 0.5, geneList = NULL)
}
\arguments{
\item{VarSelectObject}{A list object outputed by the VariableSelection function.}

\item{numGroups}{The number of groups to be made when stratifying by risk groups. Options are 2,3 and 4 (for now implementing
broader version). Default is 2.}

\item{cuts}{Numeric vector of the points in the distribution of risk scores where groups will be splitted. Needs to be of length
numGroups - 1. eg : c(0.25,0.5,0.75) when numgroups is 4. Default is 0.5.}

\item{geneList}{Optional character vector of gene names to use to generate the pie charts.}
}
\value{
ciSummary Summary of the distribution of the concordance index accross all runs.

inflPlot Bar plot of frequency of the 20 most selected features.

topHits Character vector of the top 10 most selected features.

average.risk Average predicted risk score for each patient in the data set.

data.out The data that was used for the analysis.

selectInflPlot Volcano plot of the selection frequency against the average mean coefficient of each feature accross all runs.
Note that this plot is interactive.

RiskRefit Refitted cox proportional hazard model with the predicted average risk score as the continuous covariate.

RiskHistogram Histogram of the density distribution of the average predicted risk score. Note it has been rescaled from 0 to 10
for simplicity.

Fits Data frame reporting the coefficients found for each feature at each run.

time.type Time unit used. Options are Days or Months.

RiskScoreSummary Distribution summary of the average predicted risk score.

KM_Plot Kaplan-Meier plot stratified by risk group.

SurvSum Summary of survival metrics per risk group.

mut_Plot Mutation distribution by features bar plot.

PieChart Interactive pie chart of the mutation dsitribution using either the most frequently selected features or
the manually inputted gene list. Each of the pies represent one of the risk groups.

GenesUsed Character vector of the features used to make the pie charts.
}
\description{
This functions let's the user study the output of the VariableSelection function. This function takes as input
one of the (or the one) objects returned from the different machine learning algorithms chosen previously.
Only one such object can be inputted everytime in the getResults function.
}
\examples{
load("ExampleRun_cv_LASSO.Rdata")
getResults(LASSO,numGroups=4,cuts=c(0.25,0.5,0.75),geneList=NULL)
}
\keyword{Results}
