% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllSplits.R
\name{GetSplits}
\alias{GetSplits}
\title{GetSplits}
\usage{
GetSplits(x, inames = NULL, ilevels = NULL, nodeNames = c())
}
\arguments{
\item{x}{an essemble forest object containing all the trees made.}
}
\description{
Enable the user to retrieve all the splits made in the conditional random forest from the party package. Used in the VariableSelection
function.
}
\keyword{Splits}
