% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{launchApp}
\usage{
launchApp()
}
\description{
Run OncoCast App, enabling the user to either generate an OncoCast run (though not recommended)
if it can be ran in the console. Or create a flexible framework to explore the results of an
OncoCast run that has been saved.
}
