% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutSummary.R
\name{mutSummary}
\alias{mutSummary}
\title{mutSummary}
\usage{
mutSummary(data, average.risk, topHits, numGroups, geneList)
}
\arguments{
\item{data}{The data used for the upstream analysis with an extract column indicating the risk group of each patients.
Can be obtained from the output of the riskStrat function.}

\item{numGroups}{The number of groups to be made when stratifying by risk groups. Options are 2,3 and 4 (for now implementing
broader version). Default is 2.}

\item{geneList}{Optional character vector of gene names to use to generate the pie charts. Default is NULL, which
leads to using the 5 most frequently selected features.}

\item{cuts}{Numeric vector of the points in the distribution of risk scores where groups will be splitted. Needs to be of length
numGroups - 1. eg : c(0.25,0.5,0.75) when numgroups is 4. Default is 0.5.}
}
\value{
mut_Plot Mutation distribution by features bar plot.

PieChart Interactive pie chart of the mutation dsitribution using either the most frequently selected features or
the manually inputted gene list. Each of the pies represent one of the risk groups.

GenesUsed Character vector of the features used to make the pie charts.
}
\description{
Let's the user explore the distribution of the mutational profiles by risk groups. Only works with binary predictors.
}
\examples{
library(OncoCast)
test <- OncoCast(data=survData,formula=Surv(time,status)~.,
                          method=c("LASSO"),
                          runs = 25,cores = 1,sampling = "cv",
                          pathResults = "./Test/",studyType = "ExampleRun",save=F)
out <- outputSummary(test$LASSO)
riskout <- riskStrat(survData,out$average.risk,numGroups = 2,cuts=0.5)
data <- riskout$data.out
topHits <- out$topHits
mut.results <- mutSummary(data,topHits,numGroups = 2,geneList=NULL)

}
