% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predIncomingSurv}
\alias{predIncomingSurv}
\title{predIncomingSurv}
\usage{
predIncomingSurv(OC_object, new.data, surv.print = NULL, riskRefit)
}
\arguments{
\item{OC_object}{Output of the OncoCast function.}

\item{new.data}{New data set containing the information of the incoming patients. Should be a dataframe
with patients as rows and features as columns.}

\item{surv.print}{A numeric vector indicating the patients for which the user wishes to print the predicted
survival curves.}

\item{riskRefit}{The refitted cox proportional hazard model with the risk score as predictor.}
}
\value{
data.out : The data frame inputted in the function with an additional column giving the predicted risk
score of the incoming patients.

RiskHist : A histogram of the distribution of the risk scores of patients in the given dataset.

IncKM : An interactive Kaplan-Meier plot of the selected patients in the surv.print argument.
}
\description{
This function let's user predict the genetic risk score of new/incoming patients. This function
makes use of the output of the OncoCast function and an inputed data set similar to the one use to generate
those results. The user can retrieve the genetic risk score of those new patients and their predicted survival curves.
}
\examples{
library(OncoCast)
test <- OncoCast(data=survData,formula = Surv(time,status)~.,
method = "LASSO",runs = 30,
save = FALSE,nonPenCol = NULL,cores =1)
results <- getResults_OC(OC_object=test$LASSO,data=survData,
cuts=c(0.2,0.4,0.6,0.8),
geneList=NULL,mut.data=TRUE)
new.data <- as.data.frame(matrix(rbinom(5*20,1,0.5),nrow=20,ncol = 5))
colnames(new.data) <- c("ImpCov1","ImpCov2","ImpCov3","ImpCov4","Cov7")
rownames(new.data) <- paste0("Incoming",1:20)
Incoming <- predIncomingSurv(test$LASSO,new.data=new.data,
surv.print = c(5,10,15),riskRefit = results$RiskRefit)
}
