% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appFunctions.R
\name{predIncomingSurv_app}
\alias{predIncomingSurv_app}
\title{predIncomingSurv_app}
\usage{
predIncomingSurv_app(OC_object, new.data, surv.print = NULL, riskRefit)
}
\arguments{
\item{OC_object}{OncoCast() function output}

\item{new.data}{Data to be predicted}

\item{surv.print}{Numeric vector of indexes to be predited}

\item{riskRefit}{Corresponding getResults_OC() refitted cox model}
}
\description{
predIncomingSurv_app
}
