% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appFunctions.R
\name{Progno.tab}
\alias{Progno.tab}
\title{Progno.tab}
\usage{
Progno.tab(OC_object, data, geneList = NULL)
}
\arguments{
\item{OC_object}{OncoCast() function output}

\item{data}{The corresponding data to create the OncoCast run}

\item{geneList}{A vector of genes to be annotated}
}
\description{
Progno.tab
}
