% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{server}
\alias{server}
\title{app server function}
\usage{
server(input, output)
}
\arguments{
\item{input}{input for server}

\item{output}{output for server}
}
\value{
Web-based application
}
\description{
This file contains the  server functions for running the internal application of
OncoCast.
}
\examples{
library(OncoCast)
}
\keyword{application}
