% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\docType{data}
\name{ui}
\alias{ui}
\title{app UI and server functions}
\format{
An object of class \code{shiny.tag} of length 3.
}
\usage{
ui
}
\arguments{
\item{...}{Futher arguments}
}
\value{
Web-based application
}
\description{
This file contains the ui and server functions for running the internal application of
OncoCast.
}
\examples{
library(OncoCast)
}
\keyword{application}
