% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appFunctions.R
\name{validate_app}
\alias{validate_app}
\title{validate_app}
\usage{
validate_app(OC_object, LassoFits, ori.risk, qts, in.data, formula)
}
\arguments{
\item{OC_object}{OncoCast() function output}

\item{LassoFits}{Fits outputted by getResults_OC()}

\item{ori.risk}{Predicted risk score of OC_Object}

\item{qts}{Quantiles of the cuts perforned}

\item{in.data}{Data to be used for validation}

\item{formula}{Formula to be used for validation}
}
\description{
validate_app
}
