% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipeline.R
\name{VariableSelection}
\alias{VariableSelection}
\title{VariableSelection}
\usage{
VariableSelection(data, family = "cox", formula, method = c("LASSO",
  "RIDGE", "ENET", "CF", "RF", "GBM"), runs = 100, penalizeCol = NULL,
  sampling = "cv", nTree = 1000, nsplits = 10, cores = 1,
  n.trees = 1000, interactions = 2, shrinkage = 0.1, min.node = 20,
  pathResults = "", studyType = "", save = T)
}
\arguments{
\item{data}{name of the data frame with variables as columns and patients as rows.}

\item{family}{Type of analysis to be performed depending on the outcome of interest. Options are "gaussian","cox" and"binomial".}

\item{formula}{Must match the family chosen in the previous parameter. For the "gaussian" and "binomial" families the
formula should be of the form 'y~.'. For the 'cox' family, a survival formula with the names of the variables to be used in the data frame provided in the first argument.
 eg : Surv(time,status)~. or Surv(time1,time2,status)~. (Note all the variable available will be used regardless of the right
 side of the formula).}

\item{method}{Character vector of the names of the method(s) to be used, options are : 1) LASSO ("LASSO") 2) Ridge ("RIDGE")
3) Elastic Net ("ENET") 4) Conditional inference random forest ("CF") 5) Classical random forest ("RF"). Default is all.}

\item{runs}{Number of cross validation iterations to be performed. Default is 100.}

\item{penalizeCol}{Name of variables you do not with to penalize (available only for LASSO, RIDGE and ENET methods). Default is NULL.}

\item{sampling}{The method use for sampling, options are bootstrapping ("boot") and cross-validation ("cv").
Default is cross-validation.}

\item{nTree}{Number of trees to be made at each iteration (available only for methods CF and RF). Default is 1000.}

\item{nsplits}{Maximum number of splits to be made in each tree of the RF method. Default is 10.}

\item{cores}{If you wish to run this function in parallel set the number of cores to be used to be greater than 1. Default is 1.
CAUTION : Overloading your computer can lead to serious issues, please check how many cores are available on your machine
before selecting an option !}

\item{pathResults}{String of where you wish to output the results. Default is current directory.}

\item{studyType}{String that will be the prefix to the name of the outputed results. Default is empty.}
}
\value{
CI : For each iteration the concordance index of the generated model will be calculated on the testing set

predicted : The predicted risk for all the patients in the testing set

fit : For LASSO, RIDGE and ENET methods return the refitted cox proportional hazard model with the beta coefficients found
in the penalized regression model.

Vars : For methods CF and RF, represent the relative influence of each of the variables.

Selection : The number of times each variables is selected accross all the trees. For the CF method only.

count : The number of trees that had at least one split. For the CF method only.
}
\description{
This functions let's the user select one or multiple machine learning algorithms. This is intended for survival data, and
the mehods can handle left truncated survival data. The inputed data should be a data frame with columns representing the
variables of interest while the rows should correspond to patients.
The output will vary depending on the methods. However they will all return a list of length equal to the number of
crossvalidation performed, including the predicted risk score at each cross validation for all the patients falling in the
test set.
}
\examples{
library(SurvivalVariableSelection)
}
\keyword{Selection}
