% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbmExtras.R
\name{relative.influence.noprint}
\alias{relative.influence.noprint}
\title{relative.influence.noprint}
\usage{
relative.influence.noprint(object, n.trees, scale. = FALSE,
  sort. = FALSE)
}
\description{
relative.influence.noprint
}
